@@ -1,5 +1,6 @@
-import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
+import { User } from './_models/user';
+import { AccountService } from './_services/account.service';
 
 @Component({
   selector: 'app-root',
@@ -7,18 +8,23 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent implements OnInit {
+   
     title = 'Dating apps';
-    users: any;
-  constructor(private http: HttpClient) { }
+  constructor(private accountService:AccountService) { }
 
     ngOnInit(): void { // on init vient apres constructor 
-      this.http.get('https://localhost:5001/api/users/').subscribe( {
-        next: response => this.users = response,
-        error: err => console.log(err),
-        complete: () => console.log('Request completed')
-        
-      })
+      this.setCurrentUser();
     }
+
+    
+
+    setCurrentUser(){
+      const userString = localStorage.getItem('user');
+      if(!userString) return;
+      const user: User = JSON.parse(userString);
+      this.accountService.setCurrentUser(user);
+    }
+      
 
   
 
